+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@types/pino": "^5.15.4",
     "apollo-server-express": "2.9.16",
+    "co": "^4.6.0",
     "cors": "2.8.5",
     "express": "4.17.1",
     "graphback": "0.10.2",
@@ -27,6 +28,7 @@
   "devDependencies": {
     "@types/glob": "^7.1.1",
     "@types/graphql": "14.2.3",
+    "@types/lodash": "^4.14.149",
     "@types/node-fetch": "2.5.4",
     "ts-node": "8.6.2",
     "tslint": "5.20.1",
