+++ tsconfig.json
@@ -11,7 +11,7 @@
       "es2016"
     ],
     "resolveJsonModule": true,
-    "noImplicitAny": true,
+    "noImplicitAny": false,
     "preserveConstEnums": true,
     "strict": true,
     "strictNullChecks": true,
